Cellula_2week_[MohamedBadra]/Toxic_content_classification_project/requirements.txt
fastapi
uvicorn
motor
torch
transformers
peft
pillow
requests
httpx
python-multipart
python-dotenv